requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
tushare==1.3.5
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
apscheduler==3.6.0
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.0
aiofiles==23.2.1
